+++ circom-proving/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "circom-proving"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ark-ff = { version = "0.3.0", default-features = false, features = [ "asm"] }
+ark-ec = { version = "^0.3.0", default-features = false }
+ark-relations = { version = "^0.3.0", default-features = false }
+ark-serialize = { version = "^0.3.0", default-features = false, features = [ "derive" ] }
+ark-groth16 = { git = "https://github.com/arkworks-rs/groth16", rev = "765817f", default-features = false, features = ["parallel"] }
+ark-std = { version = "^0.3.0", default-features = false }
+ark-bn254 = { version = "0.3.0" }
+ark-circom = { git = "https://github.com/vacp2p/ark-circom", branch = "wasm", default-features = false, features = ["circom-2"] }
+num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
+frame-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
+# For ark-circom
+once_cell = "1.14.0"
+cfg-if = "1.0"
+wasmer = { version = "2.3.0", default-features = false }
+num-traits = "0.2.11"
+thiserror = "1.0.0"
+color-eyre = "0.6.1"
+serde_json = "1.0.48"
+
+[features]
+default = ["wasmer/sys-default"]
+wasm = ["wasmer/js", "wasmer/std"]
