+++ primitives/src/proving/mod.rs
@@ -1,0 +1,2 @@
+pub mod circom_prover;
+pub use circom_prover::*;
